requirements.txt
pymupdf>=1.18.0
python-docx>=0.8.10
pytesseract>=0.3.8
opencv-python>=4.5.4
openai-whisper>=1.0
sentence-transformers>=2.2.0
transformers>=4.36.0
faiss-cpu>=1.7.2
torch>=1.10.0
numpy>=1.21.0
tqdm>=4.62.0
streamlit>=1.22.0
google-generativeai>=0.2.0
Pillow>=9.0.0
pandas>=1.3.0
